refactor: address ci-review findings for async rendering pipeline
- Move run_sync() from webcompy.aio to webcompy.testing._utils
- Update design.md and spec.md: report_error -> logging.error
- Update spec.md: asyncio_mode explicit markers (strict mode)
