+++ webcompy/aio/__init__.py
@@ -1,12 +1,11 @@
 from webcompy.aio._aio import AsyncWrapper, resolve_async
 from webcompy.aio._async_result import AsyncResult, AsyncState
-from webcompy.aio._utils import run_sync, sleep
+from webcompy.aio._utils import sleep
 
 __all__ = [
     "AsyncResult",
     "AsyncState",
     "AsyncWrapper",
     "resolve_async",
-    "run_sync",
     "sleep",
 ]
